% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{cutoff.val}
\alias{cutoff.val}
\title{A selector which uses an absolute cutoff}
\usage{
cutoff.val(attrs, k)
}
\arguments{
\item{attrs}{A \code{data.frame} containing ranks in the first column and attribute names as
row names. See \code{\link[FSelector]{cutoff.k}}.}

\item{k}{A numeric threshold between the min and max of the first column of \code{attrs}.}
}
\value{
Rownames of the first column of \code{attrs} that exceed \code{k}.
}
\description{
In the style of \code{\link[FSelector]{cutoff.k}}.
}
