% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{cvSLVarImpPlotOld}
\alias{cvSLVarImpPlotOld}
\title{Short description}
\usage{
cvSLVarImpPlotOld(res, feat.labs, multifold = TRUE)
}
\arguments{
\item{res}{was cvsl.var.imp.all}
}
\value{
A list of ggplots
}
\description{
Long description
}
