% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{extractScreen.CV.SuperLearner}
\alias{extractScreen.CV.SuperLearner}
\title{Short description}
\usage{
extractScreen.CV.SuperLearner(x, selector.library, weighted, verbose)
}
\arguments{
\item{verbose}{logical; passthru to selectFeaturesByMethodAndSelector}
}
\value{

}
\description{
Long description
}
