% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{factor_to_indicator}
\alias{factor_to_indicator}
\title{Utility function to create indicator variables from a factor}
\usage{
factor_to_indicator(col_name, dat)
}
\arguments{
\item{col_name}{Character string; column name in \code{dat} containing a
factor variable}

\item{dat}{\code{data.frame}}
}
\value{
A \code{data.frame} containing indicator column(s)
}
\description{
Utility function to create indicator variables from a factor
}
\examples{
factor_to_indicator("Species", iris)
}
