% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{featReorderAndFactorize}
\alias{featReorderAndFactorize}
\title{Utility factorizer function}
\usage{
featReorderAndFactorize(feat, val, featLabs, rev = TRUE)
}
\arguments{
\item{feat}{A vector of values that can be made into a factor}

\item{val}{A vector by which to order}

\item{featLabs}{A named character vector containing factor labels. Can be \code{NULL}.}

\item{rev}{Should the factor be sorted in ascending order by val? Defaults to \code{TRUE}.}
}
\value{
factorized version of \code{feat}
}
\description{
Used by CV SuperLearner feature selection visualization functions
}
