% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{ggally_custom_density}
\alias{ggally_custom_density}
\title{A custom density for a \code{GGally::ggpairs()} plot}
\usage{
ggally_custom_density(data, mapping, ...)
}
\arguments{
\item{data}{data set using}

\item{mapping}{aesthetics being used}

\item{...}{other arguments to add to \code{\link[ggplot2]{geom_density}}}
}
\value{
A \code{ggplot}
}
\description{
A custom density for a \code{GGally::ggpairs()} plot
}
