% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{groupBySelectionSet}
\alias{groupBySelectionSet}
\title{Short description}
\usage{
groupBySelectionSet(res)
}
\arguments{
\item{res}{returned from cvSLFeatureSelector()}
}
\value{

}
\description{
Long description
}
\examples{
\dontrun{
# based on example in SuperLearner package
set.seed(1)
n <- 100
p <- 20
X <- matrix(rnorm(n*p), nrow = n, ncol = p)
X <- data.frame(X)
Y <- X[, 1] + sqrt(abs(X[, 2] * X[, 3])) + X[, 2] - X[, 3] + rnorm(n)
res <- cvSLFeatureSelector(Y, X, family = gaussian(),
                           method = "method.NNLS",
                           SL.library = setNames(list(c("SL.mean", "screen.randomForest.imp"),
                                                      c("SL.mean", "screen.earth.backwardprune")),
                                                 c("random forest biggest diff mean",
                                                   "splines biggest diff mean")),
                           selector.library = data.frame(selector = c("cutoff.biggest.diff",
                                                                     "cutoff.k"),
                                                         k = c(NA, 3),
                                                         rowname = c("biggest diff", "top3"),
                                                         stringsAsFactors = FALSE) \%>\%
                                              tibble::column_to_rownames(),
                           nFolds = 3,
                           verbose = TRUE)
groupBySelectionSet(res)
}
}
