% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{is_not_homogenous}
\alias{is_not_homogenous}
\title{Assess whether \code{x} contains multiple unique values}
\usage{
is_not_homogenous(x)
}
\arguments{
\item{x}{a vector}
}
\value{
A boolean
}
\description{
Assess whether \code{x} contains multiple unique values
}
