% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{makeLibArgNumeric}
\alias{makeLibArgNumeric}
\title{Coerce a value to numeric or die trying}
\usage{
makeLibArgNumeric(x, errorFunction)
}
\arguments{
\item{x}{A value that should be coerced to numeric}

\item{errorFunction}{A function which elicits a useful error message if \code{x} cannot be
coerced to numeric}
}
\value{
\code{x} as numeric, if coercible; otherwise, an error is produced.
}
\description{
Coerce a value to numeric or die trying
}
\examples{
\dontrun{
errfxn = function(txt = "") {
  stop(paste("x must be NA, NULL, or coercible to numeric.", txt))
}
x <- "5"
x <- makeLibArgNumeric(x, errfxn)
}
}
