% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{plot_dag_proppr_data}
\alias{plot_dag_proppr_data}
\title{Plot DAG for PROPPR example}
\usage{
plot_dag_proppr_data()
}
\value{
DAG visualization
}
\description{
Plot DAG for PROPPR example
}
\examples{
\dontrun{
p <- plot_dag_proppr_data()
}
}
