% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{plot_dag_toy_data}
\alias{plot_dag_toy_data}
\title{Plot DAG for a toy example}
\usage{
plot_dag_toy_data()
}
\value{
DAG visualization
}
\description{
Plot DAG for a toy example of ...
}
\examples{
\dontrun{
p <- plot_dag_toy_data()
}
}
