% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{plot_heatmap_sim_data}
\alias{plot_heatmap_sim_data}
\title{Short description}
\usage{
plot_heatmap_sim_data(dat)
}
\arguments{
\item{dat}{A \code{data.frame} with outcome column \code{Y} and ID column \code{ID}.}
}
\value{
A ggplot
}
\description{
Longer description
}
\examples{
\dontrun{
dat <- sim_toy_data(n_obs = 200, rnd_seed = 620)
p <- plot_heatmap_sim_data(dat)
}
}
