% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{plot_splom_sim_data}
\alias{plot_splom_sim_data}
\title{Plot SPLOM for data with binary outcome}
\usage{
plot_splom_sim_data(
  dat,
  outcome_levels = 0:1,
  outcome_colors = RColorBrewer::brewer.pal(3, "RdBu")[c(1, 3)]
)
}
\arguments{
\item{dat}{A \code{data.frame} with outcome column \code{Y} and ID column \code{ID}.}

\item{outcome_levels}{Unique allowed values in the outcome column (\code{Y}).}

\item{outcome_colors}{Vector of colors with a length equal to \code{length(outcome_levels)}.}
}
\value{
SPLOM visualization (ggplot)
}
\description{
Plot SPLOM for \code{data.frame} \code{dat}.
}
\examples{
\dontrun{
dat <- sim_toy_data(n_obs = 200, rnd_seed = 620)
p <- plot_splom_sim_data(dat)
}
}
