% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{selectFeaturesByMethodAndSelector}
\alias{selectFeaturesByMethodAndSelector}
\title{Short description}
\usage{
selectFeaturesByMethodAndSelector(df, sel, propCol, verbose)
}
\arguments{
\item{verbose}{logical; passthru to selectFeaturesBySelector}
}
\value{

}
\description{
Long description
}
