% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{selectFeaturesBySelector}
\alias{selectFeaturesBySelector}
\title{Short description}
\usage{
selectFeaturesBySelector(df, sel, propCol, verbose)
}
\arguments{
\item{verbose}{logical; passthru to metaFeatSel}
}
\value{

}
\description{
Long description
}
