% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_elemstatlearn_data}
\alias{sim_elemstatlearn_data}
\title{Simulated mixture example with 2 classes.}
\usage{
sim_elemstatlearn_data(
  n_obs = 500,
  p_noise = 6,
  class_prop = 0.5,
  rnd_seed = NULL
)
}
\arguments{
\item{n_obs}{Number of observations to simulate.}

\item{p_noise}{Number of N(0,1) random noise covariate columns.}

\item{class_prop}{Proportion in the second class (labelled 1).}

\item{rnd_seed}{Random seed to be used when simulating.}
}
\value{
A \code{data.frame}
}
\description{
Simulated mixture example from ElemStatLearn. Two classes. Four + \code{p_noise} covariates.
}
\examples{
\dontrun{
dat <- sim_elemstatlearn_data(n_obs = 200, rnd_seed = 620)
testobs <- sim_elemstatlearn_data(n_obs = 1, rnd_seed = 801)
}
}
