% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_sl_data}
\alias{sim_sl_data}
\title{Simulate data as in a SuperLearner example}
\usage{
sim_sl_data(n_obs = 100, rnd_seed = NULL)
}
\arguments{
\item{n_obs}{Number of observations to simulate.}

\item{rnd_seed}{Random seed to be used when simulating.}
}
\value{
A \code{data.frame} with columns \code{ID}, \code{X1:20}, and \code{Y}.
}
\description{
Simulate data for a 20 continuous (normally-distributed) covariate, 1 continuous outcome example
where the outcome (\code{Y}) is only related to the first three covariates (\code{X1}, \code{X2},
and \code{X3}) and no covariate is related to any other covariate.
}
\examples{
\dontrun{
dat <- sim_sl_data(n_obs = 200, rnd_seed = 620)
testobs <- sim_sl_data(n_obs = 1, rnd_seed = 801)
}
}
