% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_toy_data}
\alias{sim_toy_data}
\title{Simulate data for a toy example}
\usage{
sim_toy_data(n_obs = 100, rnd_seed = NULL, verbose = FALSE)
}
\arguments{
\item{n_obs}{Number of observations to simulate.}

\item{rnd_seed}{Random seed to be used when simulating.}

\item{verbose}{Print diagnostic messages? Defaults to FALSE}
}
\value{
A \code{data.frame} with columns \code{ID}, \code{W1}, \code{W2}, \code{W3}, and
\code{Y}.
}
\description{
Simulate data for a 3 continuous covariate, 1 binary outcome example where the outcome
(\code{Y}) is unrelated to the first covariate (normally-distributed \code{W1}) but is related
to the second and third covariates (lognormally-distributed \code{W2} and \code{W3}, which are
also related to one another). Note that this function requires the \code{simcausal} package,
which may need to be installed from github like so:
\code{remotes::install_github('osofr/simcausal', build_vignettes = FALSE)}
}
\examples{
\dontrun{
dat <- sim_toy_data(n_obs = 200, rnd_seed = 620)
testobs <- sim_toy_data(n_obs = 1, rnd_seed = 801)
}
}
