% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizeScreen}
\alias{summarizeScreen}
\title{Short description}
\usage{
summarizeScreen(
  res,
  groupCols = c("method", "predictor", "screener"),
  collapseCols = NULL,
  discrete = FALSE
)
}
\arguments{
\item{res}{The \code{summary} element of the result of CV SuperLearner feature selection.}

\item{discrete}{If \code{TRUE}, instead of summarizing, simply return rows in \code{res} where
\code{discrete} is \code{TRUE}. Defaults to \code{FALSE}.}
}
\value{
A \code{data.frame}
}
\description{
Long description
}
\examples{
\dontrun{
# based on example in SuperLearner package
set.seed(1)
n <- 100
p <- 20
X <- matrix(rnorm(n*p), nrow = n, ncol = p)
X <- data.frame(X)
Y <- X[, 1] + sqrt(abs(X[, 2] * X[, 3])) + X[, 2] - X[, 3] + rnorm(n)
res <- cvSuperSelect(Y, X, gaussian(),
                     list(SL.library = setNames(list(c("SL.mean", "screen.randomForest.imp"),
                                                     c("SL.mean", "screen.earth.backwardprune")),
                                                c("random forest biggest diff mean",
                                                  "splines biggest diff mean")),
                          method = "method.NNLS",
                          selector.library = data.frame(selector = c("cutoff.biggest.diff",
                                                                     "cutoff.k"),
                                                        k = c(NA, 3),
                                                        rowname = c("biggest diff", "top3"),
                                                        stringsAsFactors = FALSE) \%>\%
                                             tibble::column_to_rownames(),
                          nFolds = 3),
                     verbose = TRUE)
summarizeScreen(res$summary, groupCols = c("method", "screener"))
}
}
