% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{cvSLVarImpPlot2}
\alias{cvSLVarImpPlot2}
\title{Short description}
\usage{
cvSLVarImpPlot2(
  res,
  featCol = "term",
  featLab = "Feature name",
  featLabs = NULL,
  valCol = "propFoldSel",
  valLab = "Proportion of CVSL folds in which feature was selected",
  catCol = "screener",
  catLab = "Screening algorithm",
  varImpTheme = theme_classic() + theme(plot.title = element_text(hjust = 0.5),
    plot.subtitle = element_text(hjust = 0.5), legend.position = "none", axis.text.x =
    element_text(angle = 60, hjust = 1, vjust = 1)),
  pointSize = 2,
  strokeSize = 1,
  maxColor = "grey10",
  title = "CV SuperLearner Variable Importance",
  subtitle = NULL,
  panelCol = NULL,
  panelLabs = NULL,
  shapeCol = NULL
)
}
\value{

}
\description{
Long description
}
\examples{
\dontrun{
# based on example in SuperLearner package
set.seed(1)
n <- 100
p <- 20
X <- matrix(rnorm(n*p), nrow = n, ncol = p)
X <- data.frame(X)
Y <- X[, 1] + sqrt(abs(X[, 2] * X[, 3])) + X[, 2] - X[, 3] + rnorm(n)
res <- cvSuperSelect(Y, X, gaussian(),
                     list(SL.library = setNames(list(c("SL.mean", "screen.randomForest.imp"),
                                                     c("SL.mean", "screen.earth.backwardprune")),
                                                c("random forest biggest diff mean",
                                                  "splines biggest diff mean")),
                          method = "method.NNLS",
                          selector.library = data.frame(selector = c("cutoff.biggest.diff",
                                                                     "cutoff.k"),
                                                        k = c(NA, 3),
                                                        rowname = c("biggest diff", "top3"),
                                                        stringsAsFactors = FALSE) \%>\%
                                             tibble::column_to_rownames(),
                          nFolds = 3),
                     verbose = TRUE)
res_sum <- summarizeScreen(res$summary, groupCols = c("method", "screener"))
cvSLVarImpPlot2(res_sum)
}
}
