% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{extractScreen.CV.SuperLearner}
\alias{extractScreen.CV.SuperLearner}
\title{Extract feature selection results from code{CV.SuperLearner} object}
\usage{
extractScreen.CV.SuperLearner(x, selector.library, weighted, verbose)
}
\arguments{
\item{x}{A named list containing objects of class \code{CV.SuperLearner}}

\item{selector.library}{A \code{data.frame} with 2 columns: "selector" and
"k". The "selector" column should contain strings naming selector functions
(typically from the \code{\link[FSelector]{FSelector}} package, with a prefix
of \code{cutoff.}). The "k" column should be the required second argument to
the named "selector" function (usually named \code{k}), if one exists. If
there is no second argument, column "k" should be set to \code{NA}. Rownames
will be utilized if they are included.}

\item{weighted}{boolean; should the estimated weights be used to weight the
feature selection?}

\item{verbose}{boolean; passed through to \link{selectFeaturesByMethodAndSelector}}
}
\value{
A list with named elements \code{whichVariable} and \code{summary}.
}
\description{
Extract feature selection results from code{CV.SuperLearner} object
}
