% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{featureSelector}
\alias{featureSelector}
\title{Short description}
\usage{
featureSelector(attrs, selector, k, env = parent.frame())
}
\arguments{
\item{env}{Environment in which \code{selector} can be found. Defaults to \code{parent.frame()}.}
}
\value{
The result of applying \code{selector} with argument \code{k} to \code{attrs}.
}
\description{
Long description
}
