% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{groupBySelectionSet}
\alias{groupBySelectionSet}
\title{Dedupe/summarize "whichVariable" element of \code{cvSLFeatureSelector} result}
\usage{
groupBySelectionSet(res)
}
\arguments{
\item{res}{\code{list} result from \link{cvSLFeatureSelector}. Expected
to contain a \code{data.frame} element named "whichVariable".}
}
\value{
The "whichVariable" \code{data.frame}, deduped/summarized by the
"keep_bin" column and with the "combo_method", "selector", and "k" columns
converted to list-columns.
}
\description{
Dedupe/summarize "whichVariable" element of \code{cvSLFeatureSelector} result
}
\examples{
\dontrun{
# based on example in SuperLearner package
dat <- sim_sl_data(n_obs = 100, rnd_seed = 1)
res <- cvSLFeatureSelector(dat \%>\% pull(Y), dat \%>\% select(-c(ID, Y)), family = gaussian(),
                           method = "method.NNLS",
                           SL.library = setNames(list(c("SL.mean", "screen.randomForest.imp"),
                                                      c("SL.mean", "screen.earth.backwardprune")),
                                                 c("random forest biggest diff mean",
                                                   "splines biggest diff mean")),
                           selector.library = data.frame(selector = c("cutoff.biggest.diff",
                                                                     "cutoff.k"),
                                                         k = c(NA, 3),
                                                         rowname = c("biggest diff", "top3"),
                                                         stringsAsFactors = FALSE) \%>\%
                                              tibble::column_to_rownames(),
                           nFolds = 3,
                           verbose = TRUE)
groupBySelectionSet(res)
}
}
