% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{metaFeatSel}
\alias{metaFeatSel}
\title{Short description}
\usage{
metaFeatSel(selector, k, x, xNames, env = parent.frame(), verbose = FALSE)
}
\arguments{
\item{env}{environment; passthru to featureSelector. defaults to parent.frame()}

\item{verbose}{logical; if TRUE, print retained features. defaults to FALSE}
}
\value{

}
\description{
Long description
}
