% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{screen.CV.SuperLearner}
\alias{screen.CV.SuperLearner}
\title{Wraps \code{cvSLFeatureSelector} for use as a \code{SuperLearner} screening algorithm}
\usage{
screen.CV.SuperLearner(Y, X, family, obsWeights, id, ...)
}
\value{

}
\description{
Long description
}
\examples{
\dontrun{
screen.CV.SuperLearner(Y, X, family = binomial(), obsWeights = NULL, id = NULL, verbose = TRUE)
}
}
